+++ lib/modules/application.dart
@@ -1,5 +1,6 @@
 export 'package:arche/src/application/bus.dart' show ArcheBus;
-export 'package:arche/src/application/config.dart' show ArcheConfig;
+export 'package:arche/src/application/config.dart'
+    show ArcheConfig, ConfigEntry;
 export 'package:arche/src/application/logger.dart' show ArcheLogger;
 export 'package:arche/src/application/translator.dart'
     show Translator, StringTranslator;