+++ lib/src/abc/kvrw.dart
@@ -1,7 +1,37 @@
+abstract mixin class BaseIO<V> {
+  void delete();
+  void write(V value);
+  V get();
+  V getOr(V other) {
+    if (has()) {
+      return other;
+    }
+    return get();
+  }
+
+  V getOrWrite(V other) {
+    if (has()) {
+      return get();
+    } else {
+      write(other);
+      return other;
+    }
+  }
+
+  V? tryGet() {
+    if (has()) {
+      return get();
+    }
+    return null;
+  }
+
+  bool has();
+}
+
 abstract mixin class KVIO<K, V> {
   Map<K, V> read();
   void write(K key, V value);
-  void delkey(K key);
+  void delete(K key);
   void writeAll(Map<K, V> m) {
     for (var e in m.entries) {
       write(e.key, e.value);